feat: Add slots for video and file upload components and alerts
This change add plugin slots for the file and video upload components, and the alerts components on those pages.

(cherry picked from commit b42f8bc571251028b38fb6e5560f41868a07774c)
(cherry picked from commit a9430982e7478004909c61f38e6af318a70e1b8e)
