+++ src/files-and-videos/generic/FileTable.jsx
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import isEmpty from 'lodash/isEmpty';
-import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
+import { useIntl } from '@edx/frontend-platform/i18n';
 import {
   CardView,
   DataTable,
@@ -42,9 +42,8 @@ const FileTable = ({
   maxFileSize,
   thumbnailPreview,
   infoModalSidebar,
-  // injected
-  intl,
 }) => {
+  const intl = useIntl();
   const pageCount = Math.ceil(files.length / 50);
   const columnSizes = {
     xs: 12,
@@ -327,8 +326,6 @@ FileTable.propTypes = {
   maxFileSize: PropTypes.number.isRequired,
   thumbnailPreview: PropTypes.func.isRequired,
   infoModalSidebar: PropTypes.func.isRequired,
-  // injected
-  intl: intlShape.isRequired,
 };
 
 FileTable.defaultProps = {
@@ -336,4 +333,4 @@ FileTable.defaultProps = {
   handleLockFile: () => {},
 };
 
-export default injectIntl(FileTable);
+export default FileTable;
