+++ src/plugin-slots/EditFileErrorAlertsSlot/index.jsx
@@ -1,0 +1,5 @@
+import { PluginSlot } from '@openedx/frontend-plugin-framework';
+
+const EditFileErrorAlertsSlot = () => <PluginSlot id="edit_file_error_alerts_slot" />;
+
+export default EditFileErrorAlertsSlot;
